requirements.txt
playwright==1.43.0
systemd-python>=235
python-systemd>=0.0.9